requirements.txt
asgiref==3.7.2
Django==5.0.2
django-extensions==3.2.3
django-livereload-server==0.5.1
django-taggit==5.0.1
Markdown==3.5.2
python-dotenv==1.0.1
sqlparse==0.4.4
tornado==6.4
